new board ID system and better output
ID is now assigned using an int that is incremented with each addition.
Boards are now fetched by Id using Linq. Going to investigate later to see if its more or less efficient than a dictionary with the Id as a key.

output now displays win, relationship, and board counts.